refactor(SaveToDrive): Move logic to Utils class
refactor: Clean code
