+++ app/src/main/java/com/infomaniak/mail/data/models/Attachment.kt
@@ -17,11 +17,8 @@
  */
 package com.infomaniak.mail.data.models
 
-import android.content.ComponentName
 import android.content.Context
-import android.content.Intent
 import androidx.annotation.DrawableRes
-import androidx.core.content.FileProvider
 import androidx.core.net.toFile
 import androidx.core.net.toUri
 import com.infomaniak.lib.core.utils.Utils.enumValueOfOrNull
@@ -66,7 +63,7 @@ class Attachment : EmbeddedRealmObject {
 
     inline val downloadUrl get() = ApiRoutes.resource(resource!!)
 
-    private inline val safeMimeType get() = if (mimeType == MIMETYPE_UNKNOWN) name.guessMimeType() else mimeType
+    inline val safeMimeType get() = if (mimeType == MIMETYPE_UNKNOWN) name.guessMimeType() else mimeType
 
     fun initLocalValues(name: String, size: Long, mimeType: String, uri: String): Attachment {
         this.name = name
@@ -106,25 +103,6 @@ class Attachment : EmbeddedRealmObject {
 
     fun getUploadLocalFile() = uploadLocalUri?.toUri()?.toFile()
 
-    fun openWithIntent(context: Context): Intent {
-        val uri = FileProvider.getUriForFile(context, context.getString(R.string.ATTACHMENTS_AUTHORITY), getCacheFile(context))
-        return Intent().apply {
-            action = Intent.ACTION_VIEW
-            flags = Intent.FLAG_GRANT_READ_URI_PERMISSION
-            setDataAndType(uri, safeMimeType)
-        }
-    }
-
-    fun saveToDriveIntent(context: Context): Intent {
-        val uri = FileProvider.getUriForFile(context, context.getString(R.string.ATTACHMENTS_AUTHORITY), getCacheFile(context))
-        return Intent().apply {
-            component = ComponentName(DRIVE_PACKAGE, DRIVE_CLASS)
-            action = Intent.ACTION_SEND
-            flags = Intent.FLAG_GRANT_READ_URI_PERMISSION
-            putExtra(Intent.EXTRA_STREAM, uri)
-        }
-    }
-
     enum class AttachmentDisposition {
         INLINE,
         ATTACHMENT,
@@ -144,10 +122,5 @@ class Attachment : EmbeddedRealmObject {
         VCARD(R.drawable.ic_file_vcard),
         VIDEO(R.drawable.ic_file_video),
         UNKNOWN(R.drawable.ic_file_unknown),
-    }
-
-    companion object {
-        private const val DRIVE_PACKAGE = "com.infomaniak.drive"
-        private const val DRIVE_CLASS = "com.infomaniak.drive.ui.SaveExternalFilesActivity"
     }
 }
