+++ app/src/main/java/com/infomaniak/mail/ui/main/thread/ThreadFragment.kt
@@ -58,9 +58,9 @@ import com.infomaniak.mail.databinding.FragmentThreadBinding
 import com.infomaniak.mail.ui.MainViewModel
 import com.infomaniak.mail.ui.alertDialogs.*
 import com.infomaniak.mail.ui.main.thread.ThreadViewModel.OpenThreadResult
-import com.infomaniak.mail.ui.main.thread.actions.DownloadAttachmentProgressDialog
 import com.infomaniak.mail.ui.newMessage.NewMessageActivityArgs
 import com.infomaniak.mail.utils.*
+import com.infomaniak.mail.utils.AttachmentIntentUtils.DOWNLOAD_ATTACHMENT_RESULT
 import com.infomaniak.mail.utils.ExternalUtils.findExternalRecipients
 import com.infomaniak.mail.utils.UiUtils.dividerDrawable
 import dagger.hilt.android.AndroidEntryPoint
@@ -271,7 +271,7 @@ class ThreadFragment : Fragment() {
     private fun shouldLoadDistantResources(): Boolean = localSettings.externalContent == ExternalContent.ALWAYS && isNotInSpam
 
     private fun observeOpenAttachment() {
-        getBackNavigationResult(DownloadAttachmentProgressDialog.DOWNLOAD_ATTACHMENT_RESULT, ::startActivity)
+        getBackNavigationResult(DOWNLOAD_ATTACHMENT_RESULT, ::startActivity)
     }
 
     private fun initAdapter() {
